pep8 fixes

+++ characters/views.py
@@ -6,11 +6,16 @@ from characters.forms import CharacterCreationForm, CharacterRemovalForm
 
 def index(request):
 
-	character_list = Character.objects.order_by('created_at')
-	if request.session.get('selected_character_id') != None:
-		selected_character = Character.objects.get(pk=request.session.get('selected_character_id'))
-		context = {'character_list': character_list,
-			'selected_character': selected_character}
+	user_id = request.session.get('user_id')
+	character_list = Character.objects.filter(user_id=user_id)
+	if request.session.get('selected_character_id') is not None:
+		selected_character = Character.objects.get(
+			pk=request.session.get('selected_character_id')
+		)
+		context = {
+			'character_list': character_list,
+			'selected_character': selected_character,
+		}
 		return render(request, 'characters/index.html', context)
 	else:
 		context = {'character_list': character_list}
@@ -32,13 +37,14 @@ def create(request):
 		form = CharacterCreationForm(request.POST)
 		if form.is_valid():
 			character = Character(
+				user_id=request.session.get('user_id'),
 				character_class_id=form.cleaned_data['character_class'],
 				name=form.cleaned_data['name'],
 				image=form.cleaned_data['image'],
 			)
 			character.save()
 			return redirect('index')
-	else:
+	elif request.method == 'GET':
 		form = CharacterCreationForm()
 		return render(request,'characters/create.html', {'form': form})
 
@@ -47,15 +53,16 @@ def delete(request, character_id):
 	character = Character.objects.get(pk=character_id)
 	if request.method == 'POST':
 		form = CharacterRemovalForm(request.POST)
-		if form.is_valid():
-			if form.cleaned_data['removal_check'] == "Delete":
+		if form.is_valid() and form.cleaned_data['removal_check'] == "Delete":
+				character.delete()
 				if character_id == request.session.get('selected_character_id'):
 					request.session["selected_character_id"] = None
-				character.delete()
 				return redirect('index')
 	form = CharacterRemovalForm()
-	context = {'character_id': character_id,
-			'form': form,
-			'character': character}
+	context = {
+		'character_id': character_id,
+		'form': form,
+		'character': character
+	}
 	return render(request,'characters/delete.html', context)
